server/requirements.txt
absl-py==0.12.0
adafruit-io==2.5.0
astroid==2.4.2
astunparse==1.6.3
backcall==0.2.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
faiss-cpu==1.7.0
fastapi==0.65.1
firebase-admin==5.0.0
flatbuffers==1.12
gast==0.4.0
google-api-core==1.28.0
google-api-python-client==2.6.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
h11==0.12.0
h5py==3.1.0
httplib2==0.19.1
idna==2.10
ipykernel==5.5.0
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.4
matplotlib==3.4.1
mccabe==0.6.1
# mkl-fft==1.3.0
# mkl-random==1.1.1
# mkl-service==2.3.0
msgpack==1.0.2
mtcnn==0.1.0
mysql-connector-python==8.0.25
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
paho-mqtt==1.5.1
pandas==1.2.4
parso==0.8.1
pexpect #@ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.10
proto-plus==1.18.1
protobuf==3.15.8
ptyprocess #@ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pyfcm==1.5.1
Pygments==2.7.4
PyJWT==2.1.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil #@ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
pyzmq==21.0.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.3
six #@ file:///tmp/build/80754af9/six_1605205327372/work
starlette==0.14.2
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
toml==0.10.2
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
uvicorn==0.13.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
pyfcm
keras==2.4.3

schedule~=1.1.0
python-dotenv~=0.17.1